requirements.txt
Django>=2.2.4<=3
djangorestframework==3.*
psycopg2>=2.8.3
requests==2.*